Only 3 gene list outputs
Deleted @temp array that held and wrote multi-field file for all information on sig genes. Can always look to original .diff file for that information. For the purpose of this script, we only need 3 gene lists all with a significant (user determined, default=2) fold change: 1) genes with expression values in both samples, 2) genes with expression values in sample1, 3) genes with expression values in sample2.